Implementacion funcionalidad subir archivos a servidor

+++ back/src/controllers/usuarios.controllers.js
@@ -1,4 +1,5 @@
 const Usuario = require("../models/usuarios");
+const fs = require("fs");
 
 //Devuelve la lista de usuarios
 const getUsuarios = async (req, res) => {
@@ -47,12 +48,31 @@ const getUsuarioById = async (req, res) => {
 //Agrega un usuario a la lista de usuarios
 const postUsuario = async (req, res) => {
   const { nombre, apellido, correo, contrasena } = req.body;
+  const imagen = req.files.imagen;
+  const nombreImagen = imagen.name;
+  const path = __dirname;
+  const pathImagen =
+    path.slice(0, 48) + "/public/images/usuarios/" + nombreImagen;
+
   try {
+    //Crear un usuario
     const nuevoUsuario = await Usuario.create({
       nombre,
       apellido,
       correo,
       contrasena,
+      imagen: nombreImagen,
+    });
+
+    //Mover la imagen al servidor
+    imagen.mv(pathImagen, (error) => {
+      if (error) {
+        res.status(500).json({
+          status: "ERROR",
+          message: "Error al guardar la imagen en el servidor.",
+          error,
+        });
+      }
     });
 
     res.status(201).json({
@@ -73,14 +93,47 @@ const postUsuario = async (req, res) => {
 const updateUsuario = async (req, res) => {
   const { id } = req.params;
   const { nombre, apellido, correo, contrasena } = req.body;
+  const imagen = req.files.imagen;
+  const nombreImagen = imagen.name;
+  const path = __dirname;
+  const pathImagen =
+    path.slice(0, 48) + "/public/images/usuarios/" + nombreImagen;
+
   try {
     const usuario = await Usuario.findByPk(id);
+
+    //Validar que si exista el usuario
+    if (!usuario)
+      return res.status(404).json({
+        status: "NOT FOUND",
+        error: "User not found.",
+      });
+
+    // Eliminar la imagen anterior si existe
+    if (usuario.imagen) {
+      fs.unlinkSync(pathImagen.replace(`${nombreImagen}`, `${usuario.imagen}`));
+    }
+
+    //Actualizar datos de usuario
     usuario.nombre = nombre;
     usuario.apellido = apellido;
     usuario.correo = correo;
     usuario.contrasena = contrasena;
+    usuario.imagen = nombreImagen;
 
     await usuario.save();
+
+    //Mover la imagen al servidor
+    imagen.mv(pathImagen, (error) => {
+      if (error) {
+        res.status(500).json({
+          status: "ERROR",
+          message: "Error al guardar la imagen en el servidor.",
+          error,
+        });
+      }
+    });
+
     res.status(200).json({
       status: "SUCCESS",
       message: "Usuario actualizado con éxito.",
@@ -99,12 +152,33 @@ const updateUsuario = async (req, res) => {
 const deleteUsuario = async (req, res) => {
   const { id } = req.params;
   try {
+    //Buscar el usuario en la base de datos
+    const usuario = await Usuario.findByPk(id);
+
+    //Validar que si exista el usuario
+    if (!usuario)
+      return res.status(404).json({
+        status: "NOT FOUND",
+        error: "User not found.",
+      });
+
+    const path = __dirname;
+    const pathImagen =
+      path.slice(0, 48) + "/public/images/usuarios/" + usuario.imagen;
+
+    //Eliminar usuario de la base de datos
     await Usuario.destroy({
       where: {
         id,
       },
     });
-    res.status(204);
+
+    //Eliminar la imagen si existe
+    if (usuario.imagen) {
+      fs.unlinkSync(pathImagen);
+    }
+
+    res.sendStatus(204);
   } catch (error) {
     return res.status(500).json({
       status: "ERROR",
